Small change to part 1 but no change in results
To tackle special case ". .." and ". ..." in 23167 and 33423 of the train in RU, reading of line in training file now splits from the right for a maximum of one
time. The if-statement that follows has been kept for blank lines
between sentences, and
for safety in the
case that either a token or a tag is missing.